Remove heights from toolbar
The height when closed was 42px, the height when open was auto,
which evaluated to 40px, causing the toolbar to jump up 2 pixels
when closed.
There's no need to have a 42px height on the outer element,
since the inner element has a 40px height.
There's no need to explicitly set the height to auto either.

Fixes #120