Limit the height during the opening animation
Prevents the toolbar from jumping up during the animation,
which is caused by the animating width not being enough to hold
all the elements.
If the fully open toolbar can't hold everything in one line,
the toolbar will still jump up after the animation is done,
but it's less jarring than before.

Fixes #223